+++ packages/terra-badge/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 3.50.0 - (April 13, 2021)
+
+* Changed
+  * Minor dependency version bump
+
 ## 3.49.0 - (April 6, 2021)
 
 * Changed